\name{HyperSBM-package}
\alias{HyperSBM-package}
\alias{HyperSBM}
\docType{package}
\title{
\packageTitle{HyperSBM}
}
\description{
\packageDescription{HyperSBM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HyperSBM}
\packageIndices{HyperSBM}
  The HyperSBM package contains functions to specify and fit Stochastic Blockmodel for hypergraphs. 
  It allows performing model-based clustering when the relationship among nodes are not limited to pairwise interactions (edges in graphs), but include higher-order interactions among three or more nodes (hyperedges).
  The main function relies on a variational implementation of the expectation-maximization algorithm to perform model parameters and latent groups estimation. 
  It also provides tools for generating hypergraphs according to a simplified latent structure (affiliation sub-model).
}
\author{
  Luca Brusa, Catherine Matias
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
