\name{HyperSBM-package}
\alias{HyperSBM-package}
\alias{HyperSBM}
\docType{package}
\title{
\packageTitle{HyperSBM}
}
\description{
\packageDescription{HyperSBM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HyperSBM}
\packageIndices{HyperSBM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HyperSBM}

Maintainer: \packageMaintainer{HyperSBM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
